requirements.txt
python-telegram-bot[all]>=22
langchain>=0.3
langgraph>=0.3
langchain-ollama>=0.3
apscheduler>=3.10
sqlalchemy>=2.0
pillow>=10.0
python-dotenv>=1.0
pyyaml>=6.0
httpx>=0.27
